fix: rename sidebar
